+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/ViewTreeRecorder.swift
@@ -7,6 +7,11 @@
 #if os(iOS)
 import UIKit
 
+internal struct RecordingResult {
+    let nodes: [Node]
+    let resources: [Resource]
+}
+
 internal struct ViewTreeRecorder {
     /// An array of enabled node recorders.
     ///
@@ -15,15 +20,21 @@ internal struct ViewTreeRecorder {
     let nodeRecorders: [NodeRecorder]
 
     /// Creates `Nodes` for given view and its subtree hierarchy.
-    func recordNodes(for anyView: UIView, in context: ViewTreeRecordingContext) -> [Node] {
+    func record(_ anyView: UIView, in context: ViewTreeRecordingContext) -> RecordingResult {
         var nodes: [Node] = []
-        recordRecursively(nodes: &nodes, view: anyView, context: context)
-        return nodes
+        var resources: [Resource] = []
+        recordRecursively(nodes: &nodes, resources: &resources, view: anyView, context: context)
+        return RecordingResult(nodes: nodes, resources: resources)
     }
 
     // MARK: - Private
 
-    private func recordRecursively(nodes: inout [Node], view: UIView, context: ViewTreeRecordingContext) {
+    private func recordRecursively(
+        nodes: inout [Node],
+        resources: inout [Resource],
+        view: UIView,
+        context: ViewTreeRecordingContext
+    ) {
         var context = context
         if let viewController = view.next as? UIViewController {
             context.viewControllerContext.parentType = .init(viewController)
@@ -37,11 +48,14 @@ internal struct ViewTreeRecorder {
         if !semantics.nodes.isEmpty {
             nodes.append(contentsOf: semantics.nodes)
         }
+        if !semantics.resources.isEmpty {
+            resources.append(contentsOf: semantics.resources)
+        }
 
         switch semantics.subtreeStrategy {
         case .record:
             for subview in view.subviews {
-                recordRecursively(nodes: &nodes, view: subview, context: context)
+                recordRecursively(nodes: &nodes, resources: &resources, view: subview, context: context)
             }
         case .ignore:
             break